+++ packages/pluggable-widgets-tools/src/typings-generator/generate.ts
@@ -25,6 +25,7 @@ const importableModules = [
         "ListWidgetValue",
         "NativeIcon",
         "NativeImage",
+        "ObjectItem",
         "Option",
         "ReferenceSetValue",
         "ReferenceValue",